added component bookCardList
